Add media query to buttons
This is a small update that attempts to add some responsive design to the page by making the Rock, Paper and Scissors buttons wrap when the page dips below a certain a width.

The other issue with making the page responsive is that the game log (the CPU's responses) was not wrapping, causing the whole box to stretch out unnaturally past a certain width. Enabling text wrapping fixes this but breaks the typing illusion, due to how the animation behind it works.

For now, it's more important to me that the page elements don't warp unnaturally on narrow screens, but maybe I'll look into adapting the typing animation for the media query size as well.